% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eisa-package.R
\docType{package}
\name{eisa-package}
\alias{eisa}
\alias{eisa-package}
\title{eisa: Exon-Intron Split Anaalysis (EISA)}
\description{
Exon-intron split analysis (EISA) uses ordinary RNA-seq data to measure
    changes in mature RNA and pre-mRNA reads across different experimental conditions
    to quantify transcriptional and post-transcriptional regulation of gene expression.
    For details see Gaidatzis et al., Nat Biotechnol 2015. doi: 10.1038/nbt.3269.
}
\author{
\strong{Maintainer}: Michael Stadler \email{michael.stadler@fmi.ch}

Authors:
\itemize{
  \item Dimos Gaidatzis \email{dimosthenis.gaidatzis@fmi.ch}
  \item Lukas Burger \email{lukas.burger@fmi.ch}
}

}
\keyword{internal}
