% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eisa-package.R
\docType{package}
\name{eisaR-package}
\alias{eisaR}
\alias{eisaR-package}
\title{eisaR: Exon-Intron Split Analysis (EISA) in R}
\description{
Exon-intron split analysis (EISA) uses ordinary RNA-seq data to measure changes in mature RNA and pre-mRNA reads across different experimental conditions to quantify transcriptional and post-transcriptional regulation of gene expression. For details see Gaidatzis et al., Nat Biotechnol 2015. doi: 10.1038/nbt.3269. eisaR implements the major steps of EISA in R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fmicompbio/eisaR}
  \item Report bugs at \url{https://github.com/fmicompbio/eisaR/issues}
}

}
\author{
\strong{Maintainer}: Michael Stadler \email{michael.stadler@fmi.ch}

Authors:
\itemize{
  \item Michael Stadler \email{michael.stadler@fmi.ch}
  \item Dimos Gaidatzis \email{dimosthenis.gaidatzis@fmi.ch}
  \item Lukas Burger \email{lukas.burger@fmi.ch}
  \item Charlotte Soneson \email{charlotte.soneson@fmi.ch}
}

}
\keyword{internal}
