% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractTxSeqs.R
\name{extractTxSeqs}
\alias{extractTxSeqs}
\title{Extract (spliced or unspliced) transcript sequences}
\usage{
extractTxSeqs(gtf, genome, type = c("spliced", "unspliced"), verbose = TRUE)
}
\arguments{
\item{gtf}{The path to a gtf file}

\item{genome}{A \code{DNAStringSet} object with the genome sequence}

\item{type}{Either 'spliced' or 'unspliced'}

\item{verbose}{Logical, whether or not to print messages}
}
\value{
A \code{DNAStringSet} object with intronic sequences
}
\description{
Extract (spliced or unspliced) transcript sequences
}
\author{
Charlotte Soneson
}
