% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractIntronSeqs.R
\name{extractIntronSeqs}
\alias{extractIntronSeqs}
\title{Extract intron sequences}
\usage{
extractIntronSeqs(
  gtf,
  genome,
  type = c("collapse", "separate"),
  flanklength = 90L,
  joinOverlappingIntrons = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{gtf}{The path to a gtf file}

\item{genome}{A \code{DNAStringSet} object with the genome sequence}

\item{type}{Either 'collapse' or 'separate'}

\item{flanklength}{The length of the exonic/genomic flanking sequence that
is added to each intron}

\item{joinOverlappingIntrons}{Whether to join introns of the same gene if
they overlap after expanding with the flanking sequences}

\item{verbose}{Logical, whether or not to print messages}
}
\value{
A \code{DNAStringSet} object with intronic sequences
}
\description{
Extract intron sequences
}
\author{
Charlotte Soneson
}
