% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractIntronSeqs.R
\name{extractIntronSeqs}
\alias{extractIntronSeqs}
\title{Extract intron sequences}
\usage{
extractIntronSeqs(gtf, genome, type = "collapse", flanklength = 90)
}
\arguments{
\item{gtf}{The path to a gtf file}

\item{genome}{A \code{DNAStringSet} object with the genome sequence}

\item{type}{Either 'collapse' or 'separate'}

\item{flanklength}{The length of the exonic flanking sequence}
}
\value{
A \code{DNAStringSet} object with intronic sequences
}
\description{
Extract intron sequences
}
