% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractTxSeqs.R
\name{extractTxSeqs}
\alias{extractTxSeqs}
\title{Extract (spliced or unspliced) transcript sequences}
\usage{
extractTxSeqs(gtf, genome, type = "spliced")
}
\arguments{
\item{gtf}{The path to a gtf file}

\item{genome}{A \code{DNAStringSet} object with the genome sequence}

\item{type}{Either 'spliced' or 'unspliced'}
}
\value{
A \code{DNAStringSet} object with intronic sequences
}
\description{
Extract (spliced or unspliced) transcript sequences
}
\author{
Charlotte Soneson
}
