% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runEISA.R
\name{runEISA}
\alias{runEISA}
\title{Run Exon-Intron Split Analaysis.}
\usage{
runEISA(cntEx, cntIn, cond, method = c("published", "new"), ...)
}
\arguments{
\item{cntEx}{Gene by sample \code{matrix} with exonic counts.}

\item{cntIn}{Gene by sample \code{matrix} with intronic counts. Must have the
same structure as \code{cntEx} (same number and order of rows and columns).}

\item{cond}{\code{numeric}, \code{character} or \code{factor} with two levels
that groups the samples (columns of \code{cntEx} and \code{cntIn} into two
conditions. The contrast will be defined as secondLevel - firstLevel.}

\item{method}{One of \code{"published"} or \code{"new"}. If \code{"published"}
(the default), sample normalization, gene filtering and calculation of
contrasts is performed as described in Gaidatzis et al. 2015, and the
statistical analysis is based on \code{\link[edgeR]{glmFit}} and \code{\link[edgeR]{glmLRT}}.
If \code{method="new"}, normalization will be performed using TMM as
implemented in \code{\link[edgeR]{calcNormFactors}}, the contrast are
calculated using \code{\link[edgeR]{predFC}}, and the statistical analysis
will use the quasi-likelyhood framework implemented in \code{\link[edgeR]{glmQLFit}}
and \code{\link[edgeR]{glmQLFTest}}. The latter is often less stringent when
selecting quantifyable genes, but more stringent wenn calling significant
changes (especially with low numbers of replicates).}

\item{...}{additional arguments passed to the \code{\link[edgeR]{DGEList}} constructor,
such as \code{lib.size} or \code{genes}.}
}
\value{
a \code{list} with elements \describe{
\item{fracIn}{fraction intronic counts in each sample}
\item{quantGenes}{names of quantifyable genes indluded in the analysis (from rownames of \code{cntEx})}
\item{contrastName}{contrast name}
\item{contrasts}{contrast matrix for quantifyable genes, with averge log2
fold-changes in exons (\code{Dex}), in introns (\code{Din}), and averge
difference between log2 fold-changes in exons and introns (\code{Dex.Din})}
\item{DGEList}{\code{\link[edgeR]{DGEList}} object used in model fitting}
\item{tab.cond}{statistical results for differential expression between conditions,
based on both exonic and intronic counts}
\item{tab.ExIn}{statisical results for differential changes between exonic and
intronic contrast, an indication for post-transcriptional regulation.}
\item{method}{the method that was used to run EISA}
}
}
\description{
Starting from count tables with exonic and intronic counts
for two conditions, perform all the steps in EISA (normalization, identify
quantifyable genes, calculate contrasts and their significance).
}
\examples{
cntEx <- readRDS(system.file("extdata", "Fig3abc_GSE33252_rawcounts_exonic.rds",
                             package = "eisaR"))[,-1]
cntIn <- readRDS(system.file("extdata", "Fig3abc_GSE33252_rawcounts_intronic.rds",
                             package = "eisaR"))[,-1]
cond <- factor(c("ES","ES","TN","TN"))
res <- runEISA(cntEx, cntIn, cond)
plotEISA(res)

}
\references{
Analysis of intronic and exonic reads in RNA-seq data characterizes
transcriptional and post-transcriptional regulation.
Dimos Gaidatzis, Lukas Burger, Maria Florescu and Michael B. Stadler
Nature Biotechnology, 2015 Jul;33(7):722-9. doi: 10.1038/nbt.3269.
}
\seealso{
\code{\link[edgeR]{DGEList}} for \code{DGEList} object construction,
\code{\link[edgeR]{calcNormFactors}} for normalization,
\code{\link[edgeR]{glmFit}} for statisical analysis under \code{method = "published"}
and \code{\link[edgeR]{glmQLFit}} statistical analysis under \code{method = "new"}.
}
\author{
Michael Stadler
}
